Bug Fixes and Refactors
This commit includes several bug fixes and code refactors. Specifically:

- Resolved a bug in the `_lock` function of the `VoteEscrowLocker` contract where account balances were not being set correctly.
- Fixed a bug in the `lock` function of the `VoteEscrowToken` contract by removing the redundant input parameter `account`.
Implemented the `initialize` function in the `LiquidityGaugePoo`l contract that can only be called by the owner.
Redeployed the `veNPM` and `LiquidityGaugePool` contracts after addressing the aforementioned bugs.
Updated the deployment file to reflect the changes in the contract code.
Added a new ABI named `IProtocol.json` to the codebase.